back/worker/requirements.txt
bullmq==2.19.4
python-dotenv==1.2.1
minio==7.2.20
numpy==2.2.6
Pillow==11.3.0
opencv-python-headless==4.13.0.92
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
redis==6.4.0
structlog==25.5.0
watchfiles==1.1.1
asyncio==4.0.0
